requirements.txt
2to3>=1.0
alembic>=1.0.3
aniso8601>=4.0.1
asn1crypto>=0.24.0
astroid>=2.0.4
bcrypt>=3.1.7
certifi>=2018.10.15
cffi>=1.14.0
chardet>=3.0.4
Click>=7.0
clickclick>=1.2.2
colorama>=0.4.0
connexion>=2.0.1
cryptography>=2.8
Flask>=1.0.2
Flask-JWT-Extended>=3.13.1
flask-marshmallow>=0.9.0
Flask-Migrate>=2.3.0
Flask-MySQL>=1.4.0
Flask-RESTful>=0.3.6
Flask-SQLAlchemy>=2.3.2
idna>=2.7
inflection>=0.3.1
isort>=4.3.4
itsdangerous>=1.1.0
Jinja2>=2.10
jsonschema>=2.6.0
lazy-object-proxy>=1.3.1
Mako>=1.1.2
MarkupSafe>=1.1.0
marshmallow>=2.16.3
marshmallow-enum>=1.4.1
marshmallow-sqlalchemy>=0.15.0
mccabe>=0.6.1

Naked>=0.1.31
openapi-spec-validator>=0.2.4
paramiko>=2.7.1
pycparser>=2.19
pycryptodome>=3.9.7
PyJWT>=1.4.2
pylint>=2.1.1
pylint-flask>=0.5
pylint-plugin-utils>=0.4
PyMySQL>=0.9.2
PyNaCl>=1.3.0
pyodbc>=4.0.30
python-dateutil>=2.7.5
python-editor>=1.0.3
pytz>=2018.7
PyYAML>=3.13
requests>=2.20.1
shellescape>=3.8.1
six>=1.11.0
SQLAlchemy>=1.3.15
urllib3>=1.24.1
webargs>=6.0.0
Werkzeug>=0.14.1
wfastcgi>=3.0.0
wrapt>=1.10.11
mysqlclient==1.4.6